Fixing bugs, fixing doctests
